replace scrape.py, add functions.py, add WebDriver class to prevent unnecessary repeating code, simply scraper functions
